Replace zero.dns0.eu with default dns0.eu as Secure DNS
zero.dns0.eu is too aggressive, that a lot of DNS and VPN domains got
blocked by it, the purpose of it is actually different than the Secure
DNS we expected here.

The title from its page is:

> hardened security for highly sensitive environments

So just replace it with the default dns0.eu, will be better.